eval/requirements.txt
pytest>=8.2.0
pyyaml>=6.0.1
httpx>=0.27.0
